UIXCompatibility

+++ headers/_UXCollectionView.h
@@ -4,9 +4,8 @@
 //     class-dump is Copyright (C) 1997-1998, 2000-2001, 2004-2013 by Steve Nygard.
 //
 
-#import <UXKit/UXCollectionView.h>
-
-#import "_UXCollectionViewOverdraw.h"
+#import "UXCollectionView.h"
+#import "_UXCollectionViewOverdraw-Protocol.h"
 
 @class NSString;
 
@@ -20,7 +19,7 @@
 // Remaining properties
 @property(readonly, copy) NSString *debugDescription;
 @property(readonly, copy) NSString *description;
-@property(readonly) unsigned long long hash;
+@property(readonly) NSUInteger hash;
 @property(readonly) Class superclass;
 
 @end
